Changes to Controllers/Services
Controller and Service methods were changed to avoid errors. Added some Swagger documentation etc...

+++ src/main/java/com/example/demo/Services/PollService.java
@@ -19,29 +19,43 @@ public class PollService {
         this.pollRepo = pollRepo;
     }
 
+    /* Returns a List of Polls from the DB */
     public List<Poll> getAll() {
         return pollRepo.findAll();
     }
 
+    /* Returns an Optional containing the found Poll with the given ID or an Empty one */
     public Optional<Poll> getPoll(Long id) {
         return pollRepo.findById(id);
     }
 
+    /* Adds a Poll to the DB */
     public void addPoll(Poll poll) {
         pollRepo.save(poll);
     }
 
-    public void deletePoll(Long id) {
-        pollRepo.deleteById(id);
+    /* Deletes a Poll from the DB with given ID */
+    public boolean deletePoll(Long id) {
+        Optional<Poll> pollById = pollRepo.findById(id); // we search for the Poll with the given ID
+        if (pollById.isPresent()) { // if we found it
+            pollRepo.deleteById(id); // we delete it
+            return true;
+        }
+        return false; // if there was no Poll with the given ID we return false
     }
 
-    public void updateInfo(Long id, Poll poll) {
-        Poll updatedPoll = pollRepo.getById(id);
-        updateInfo(updatedPoll, poll);
-        pollRepo.save(updatedPoll);
+    /* Updates a Poll with the given ID */
+    public Optional<Poll> updatePoll(Long id, Poll newPoll) {
+        Optional<Poll> updatedPoll = pollRepo.findById(id); // we search for the Poll with the given ID
+        if (updatedPoll.isPresent()) { // if we found it
+            updatePollInfo(updatedPoll.get(), newPoll); // we update its info
+            pollRepo.save(updatedPoll.get()); // and we save it in the DB
+        }
+        return updatedPoll; // we return an Optional containing the updated Poll or an Empty one
     }
 
-    private void updateInfo(@NotNull Poll originalPoll, @NotNull Poll newPoll) {
+    /* Helper Method that updates a Poll */
+    private void updatePollInfo(@NotNull Poll originalPoll, @NotNull Poll newPoll) {
         originalPoll.setTopic_id(newPoll.getTopic_id());
         originalPoll.setDateOfCreation(newPoll.getDateOfCreation());
         originalPoll.setDateOfClosing(newPoll.getDateOfClosing());
